backup_20250627_152302/requirements.txt
streamlit
langchain
langchain-openai
langchain-experimental
sqlalchemy
pandas
python-dotenv
pyodbc